requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2>=2.10
MarkupSafe==1.0
Werkzeug==0.15.3
biopython~=1.76
